+++ packages/yoshi/test/lint.spec.js
@@ -68,27 +68,49 @@ describe('Aggregator: Lint', () => {
       expect(res.stdout).to.contain('Missing radix parameter');
     });
 
-    it('should support a list of files to run lint on (not necessarily on tsconfig)', () => {
-      const res = test
-        .setup({
-          'app/a.ts': `parseInt("1");`,
-          'other-dir/b.tsx': `parseInt("1");`,
-          'app/dontrunonme.js': `parseInt("1");`,
-          'package.json': fx.packageJson(),
-          'tsconfig.json': fx.tsconfig({ include: ['app/a.ts'] }),
-          'tslint.json': fx.tslint({ radix: true }),
-        })
-        .execute('lint', ['app/a.ts', 'other-dir/b.tsx'], insideTeamCity);
+    describe('when file paths supplied', () => {
+      it('should work as expected if all files are within the files specified in tsconfig.json', () => {
+        const res = test
+          .setup({
+            'app/a.ts': `parseInt("1");`,
+            'other-dir/b.tsx': `parseInt("1");`,
+            'app/dontrunonme.ts': `parseInt("1");`,
+            'package.json': fx.packageJson(),
+            'tsconfig.json': fx.tsconfig({
+              include: ['app/a.ts', 'other-dir/b.tsx'],
+            }),
+            'tslint.json': fx.tslint({ radix: true }),
+          })
+          .execute('lint', ['app/a.ts', 'other-dir/b.tsx'], insideTeamCity);
 
-      expect(res.code).to.equal(1);
-      expect(res.stdout).to.contain('app/a.ts:1:1');
-      expect(res.stdout).to.contain('Missing radix parameter');
-      expect(res.stdout).to.contain('other-dir/b.tsx:1:1');
-      expect(res.stderr).to.contain('tslint exited with 2 errors');
+        expect(res.code).to.equal(1);
+        expect(res.stdout).to.contain('app/a.ts:1:1');
+        expect(res.stdout).to.contain('Missing radix parameter');
+        expect(res.stdout).to.contain('other-dir/b.tsx:1:1');
+        expect(res.stderr).to.contain('tslint exited with 2 errors');
+        expect(res.stdout).to.not.contain('app/dontrunonme.ts');
+      });
+
+      it('should log a warning if there are some files which are not specified in tsconfig.json', () => {
+        const res = test
+          .setup({
+            'app/a.ts': `console.log();`,
+            'other-dir/b.tsx': `parseInt("1");`,
+            'package.json': fx.packageJson(),
+            'tsconfig.json': fx.tsconfig({ include: ['app/a.ts'] }),
+            'tslint.json': fx.tslint({ radix: true }),
+          })
+          .execute('lint', ['app/a.ts', 'other-dir/b.tsx'], insideTeamCity);
+
+        expect(res.code).to.equal(0);
+        expect(res.stderr).to.contain(
+          '● Warning: The following files were supplied to "yoshi lint" as a pattern',
+        );
+      });
     });
   });
 
-  describe('yoshi-eslint', () => {
+  describe('eslint', () => {
     function setup(data) {
       return test.setup(
         Object.assign(
@@ -101,7 +123,7 @@ describe('Aggregator: Lint', () => {
       );
     }
 
-    it('should use yoshi-eslint', () => {
+    it('should use eslint and pass for a valid file', () => {
       const res = setup({ 'app/a.js': `parseInt("1", 10);` }).execute(
         'lint',
         [],
@@ -110,19 +132,25 @@ describe('Aggregator: Lint', () => {
       expect(res.code).to.equal(0);
     });
 
-    it('should fail with exit code 1', () => {
-      const res = setup({ 'app/a.js': `parseInt("1");` }).execute(
-        'lint',
-        [],
-        insideTeamCity,
-      );
+    it('should log warnings to the console when there are only warnings', () => {
+      const res = setup({
+        'app/a.js': `/*eslint radix: 1*/\n parseInt("1");`,
+      }).execute('lint', [], insideTeamCity);
+      expect(res.code).to.equal(0);
+      expect(res.stderr).to.contain('warning  Missing radix parameter  radix');
+    });
+
+    it('should fail with exit code 1 for an invalid file', () => {
+      const res = setup({
+        'app/a.js': `parseInt("1");`,
+      }).execute('lint', [], insideTeamCity);
       expect(res.code).to.equal(1);
       expect(res.stderr).to.contain(
         '1:1  error  Missing radix parameter  radix',
       );
     });
 
-    it('should fix linting errors and exit with exit code 0 if there are only fixable errors', () => {
+    it('should output fixes in case a "fix" option is passed', () => {
       const res = setup({
         'app/a.js':
           '/*eslint no-regex-spaces: "error"*/\nnew RegExp("foo  bar");',
@@ -160,7 +188,7 @@ describe('Aggregator: Lint', () => {
     });
   });
 
-  describe('yoshi-stylelint', () => {
+  describe('stylelint', () => {
     it('should use yoshi-stylelint', () => {
       const goodStyle = `
 p {